+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased (????-??-??)
 
-...
+* fix: abort controller behaviour with `sse-parse` used by `ApplicationLogsStream` (Node.js < 18.16)
 
 ## 8.0.1 (2023-10-17)